+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Enhancement:** Display referred visitors via UTM parameters in the `Referrals` report.
 - **Fix:** Fixed Decimal Formatting
 - **Fix**: Fixed visitors number aggregation issue.
+- **Fix:** Resolved posts view column issue.
 
 = 14.13.4 - 2025-04-29 =
 - **Enhancement:** Enforced capability check in optionUpdater.